backendApi/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
crispy-bootstrap4==2022.1
Django==4.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.0
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
